+++ Changelog.md
@@ -1,5 +1,8 @@
 ## 0.0.59
 + Fixed JSON error when creating chat completions with `muna.Beta.OpenAI.Chat.Completions.Create` method.
++ Removed `Predictor.card` field.
++ Removed `Predictor.license` field.
++ Removed `Predictor.media` field.
 
 ## 0.0.58
 + Fixed null reference exception when using `muna.Beta.OpenAI.Chat.Completions.Create` method.